Edit agreement

+++ app/Http/Controllers/AgreementController.php
@@ -42,13 +42,15 @@ class AgreementController extends Controller
     {
         $request->validate([
             'agreement_no' => 'required',
-            'date' => 'required',
+            'agreement_date' => 'required',
             'start_date' => 'required',
             'end_date' => 'required',
             'customer_id' => 'required',
+            'currency' => 'required',
         ]);
-        $data = $request->all();
-        $agreement = Agreement::create($request->except('payment_schedule'));
+        $data = $request->except('payment_schedule')+['amount' => $request->agreement_amount];
+        $data['attachments'] = json_encode($request->attachments);
+        $agreement = Agreement::create($data);
         foreach($request->payment_schedule as $key => $value){
             unset($value['id']);
             $schedule = new PaymentSchedule([
@@ -58,7 +60,7 @@ class AgreementController extends Controller
                 'status' => 'Pending',
                 'percentage' => $value['percentage'],
                 'short_description' => $value['short_description'],
-                'currency' => 'USD',
+                'currency' => $value['currency'],
             ]);
             $schedule->save();
             // $value['agreement_no'] = $request->agreement_no;
@@ -96,17 +98,40 @@ class AgreementController extends Controller
      */
     public function update(Request $request, Agreement $agreement)
     {
-        //
+        $data = $request->except('payment_schedule')+['amount' => $request->agreement_amount];
+        $data['attachments'] = json_encode($request->attachments);
+        $agreement->update($data);
+        //dd($agreement->paymentSchedules->pluck('id'));
+        PaymentSchedule::where('agreement_no','=',$agreement->agreement_no)->delete();
+        foreach($request->payment_schedule as $key => $value){
+            unset($value['id']);
+            $schedule = new PaymentSchedule([
+                'agreement_no' => $request->agreement_no,
+                'due_date' => $value['due_date'],
+                'amount' => $value['amount'],
+                'status' => $value['status'],
+                'percentage' => $value['percentage'],
+                'short_description' => $value['short_description'],
+                'currency' => $value['currency'],
+            ]);
+            $schedule->save();
+            // $value['agreement_no'] = $request->agreement_no;
+        }
+        return to_route('agreements.index');
     }
     /**
      * Upload the specified resource into storage.
      */
     public function upload(Request $request)
     {
-        if($request->has('agreement_doc_old')){
-            Storage::disk('public')->delete('agreements/'.basename($request->agreement_doc_old));
+        if($request->has('agreement_doc')){
+            if($request->has('agreement_doc_old')){
+                Storage::disk('public')->delete('agreements/'.basename($request->agreement_doc_old));
+            }
+            return Storage::url($request->file('agreement_doc')->storePublicly('agreements','public'));
+        }else if($request->has('attachments')){
+            return Storage::url($request->file('attachments')->storePublicly('attachments','public'));
         }
-        return Storage::url($request->file('agreement_doc')->storePublicly('agreements','public'));
     }
 
     /**
